Move most remaining code for wallet support of ephemeral addresses into `zcash_client_sqlite::wallet::transparent::ephemeral`.
